Repsonding to backend code review
